+++ src/Server/Persistence/Entities/EntityProcessor.cs
@@ -58,6 +58,7 @@ public sealed class EntityProcessor
     private const int IndexBbSizeX = IndexBbPosZ + 1;
     private const int IndexBbSizeY = IndexBbSizeX + 1;
     private const int IndexBbSizeZ = IndexBbSizeY + 1;
+    private const int IndexShadowRadius = IndexBbSizeZ + 1;
     private readonly IEntityFactory entityFactory;
     private readonly ILogger<EntityProcessor> logger;
     private readonly EntityMapper mapper;
@@ -116,6 +117,7 @@ public sealed class EntityProcessor
         ProcessPointLightSource(reader, builder);
         ProcessPhysics(reader, builder);
         ProcessBoundingBox(reader, builder);
+        ProcessCastShadows(reader, builder);
 
         /* Complete the entity. */
         builder.Build();
@@ -325,6 +327,20 @@ public sealed class EntityProcessor
         {
             Position = GetVector3(reader, IndexBbPosX, IndexBbPosY, IndexBbPosZ),
             Size = GetVector3(reader, IndexBbSizeX, IndexBbSizeY, IndexBbSizeZ)
+        });
+    }
+
+    /// <summary>
+    ///     Processes the CastsShadow component.
+    /// </summary>
+    /// <param name="reader">Reader.</param>
+    /// <param name="builder">Builder.</param>
+    private void ProcessCastShadows(IDataReader reader, IEntityBuilder builder)
+    {
+        if (reader.IsDBNull(IndexShadowRadius)) return;
+        builder.CastShadows(new Shadow
+        {
+            Radius = reader.GetFloat(IndexShadowRadius)
         });
     }